pyproject.toml
[project]
name = "cryptogame"
version = "0.0.7"
authors = [
  { name="UMass 2023 TA Group 5", email="hojaeson@umass.com" },
]
description = "A small example package"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "pycryptodome",
    "prettytable"
]

[project.urls]
Homepage = "https://github.com/pypa/sampleproject"
Issues = "https://github.com/pypa/sampleproject/issues"